backend_python/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
aiortc==1.6.0
numpy==1.26.2
scipy==1.11.4
webrtcvad==2.0.10
openai==1.3.7
elevenlabs==0.2.27
python-dotenv==1.0.0
pydantic==2.5.0
aiofiles==23.2.1
resampy==0.4.2
httpx==0.25.2
requests==2.31.0

